write MEDownloadTask in separate file, write some documentation
