+++ MEValleyLoader/Models/MEDownloadRequest.swift
@@ -8,16 +8,18 @@
 
 import Foundation
 
+/**
+ `MEDownloadRequest` act as initiators to download a URL, returned to user so he can cancel request if needed without knowing extra details
+ */
 public class MEDownloadRequest {
     
     public typealias SuccessCallback = (Data) -> Void
     public typealias FailureCallback = (Error) -> Void
     
-    
     var successCompletion: SuccessCallback?
     var failureCompletion: FailureCallback?
     
-    weak var task: MEDownloadTask?
+    private weak var task: MEDownloadTask?
     
     init(task: MEDownloadTask) {
         self.task = task
@@ -39,58 +41,4 @@ public class MEDownloadRequest {
         failureCompletion = completion
         return self
     }
-}
-
-public class MEDownloadTask: NSObject {
-    
-    // MARK:- Properties
-    private weak var manager: MEDownloader?
-    
-    public var url: URL
-    var data = Data()
-    var downloadTask: URLSessionDataTask?
-    private var requests = [MEDownloadRequest]() {
-        didSet {
-            if requests.isEmpty {
-                downloadTask?.cancel()
-                self.manager?.removeTask(task: self)
-            }
-        }
-    }
-    
-    // MARK:- Init
-    public init(url: URL, manager:MEDownloader) {
-        self.url = url
-        self.manager = manager
-        self.downloadTask = manager.session.dataTask(with: url)
-    }
-    
-    // MARK:- Methods
-    // MARK: Public Methods
-    public func start() {
-        downloadTask?.resume()
-    }
-    
-    public func addRequest(request: MEDownloadRequest) {
-        requests.append(request)
-    }
-    
-    public func removeRequest(request: MEDownloadRequest) {
-        if let index = (requests.firstIndex{ $0 === request }) {
-            requests.remove(at: index)
-        }
-    }
-    
-    public func taskSuccessed(data: Data) {
-        for request in requests {
-            request.successCompletion?(data)
-        }
-    }
-    
-    public func taskFailed(error: Error) {
-        for request in requests {
-            request.failureCompletion?(error)
-        }
-    }
-    
 }
